api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
mysql-connector-python==9.0.0
pyjwt==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
python-multipart==0.0.9
